i added another trigger and boxes
Sorry i messed with the stage, I wanted to raise the finishing platform so the doesnt just make it to the top that easy. Also i put the boxes in the way. I have to adjust some things and just make an objective system instead of just the triggers so...